feat(cbtop): add headless mode for CI/CD and AI agent integration
Add --headless flag and bench subcommand for programmatic benchmarking
without TTY. Enables AI coding assistants (like Claude Code) to run
cbtop benchmarks and process JSON output.

Features:
- --headless flag for basic headless mode
- cbtop bench subcommand with full options
- JSON and text output formats
- Regression detection against baseline
- Backend comparison (--compare)
- Output to file (-o/--output)

Specification:
- Add §30 to compute-block-tui-cbtop.md
- Requirements HL-001 through HL-007
- Falsification criteria F601-F605

Book:
- Add headless mode documentation to ecosystem/cbtop.md

Fixes clippy warnings:
- Replace wildcard match patterns with explicit variants
- Remove unnecessary u64 cast

Refs CBTOP-HEADLESS-001

+++ crates/cbtop/src/app.rs
@@ -705,7 +705,7 @@ impl CbtopApp {
                                 let result = unsafe { libc::statvfs(c_path.as_ptr(), stat.as_mut_ptr()) };
                                 if result == 0 {
                                     let stat = unsafe { stat.assume_init() };
-                                    let block_size = stat.f_frsize as u64;
+                                    let block_size = stat.f_frsize;
                                     let total = stat.f_blocks * block_size;
                                     let available = stat.f_bavail * block_size;
                                     let used = total.saturating_sub(available);
